backend/requirements.txt
fastapi
uvicorn
motor
pyjwt
passlib[bcrypt]
pydantic
pydantic-settings
email-validator
python-multipart
tensorflow
pillow
numpy
